cheetah-compile now takes a new argument, '-w', which write the str() of the compiled servlets to .html files.  This is for easily creating static sites.
The way -d and -R work has also been changed, so that they can be used
in conjunction with -p and -w.  Instance variables hold the value of
the various options, and everything is handled through .processFile()
and .compileFile().

-p now works just like the normal case, except it prints to stdout
instead of the .py file.  It's implemented in .compileFile().

At some point .compileDir() and .recursiveCompile() should be
removed.