Correct a bug introduced in v2.1.0 that affected odd inline imports
Added a regression test as well to prevent it from happening again, what basically
was going on is that in v2.1.0 I introduced code to properly place inline imports, for
syntax like the following:

	#try
		#import cjson
	#except ImportError
		#import simplejson
	#end try

The actual bug was with blocks of code like this:

	#def function
		#if $something
			#from package import module
			$module.do_thing()
		#end if
	#end def

Which would result in:

	def function(self):
		if something:
			import module
			module.do_thing()

Instead of the proper "from package import module" syntax.

Sorted, all tests passing now.

Signed-off-by: R. Tyler Ballance <tyler@slide.com>