remove unnecessary encoding in the default filters, now that everything is a unicode object internally
There is no longer any value in having the filters return encoded str objects,
since they need to be decoded back to unicode before they can be combined with
the rest of the template text (which is managed as a unicode object).  This
patch maintains API compatibility but fundamentally changes the behavior of
some of the filters.  RawOrEncodedUnicode and EncodeUnicode should be
deprecated and removed from the API.

Signed-off-by: R. Tyler Ballance <tyler@monkeypox.org>